refactor codes to de-clutter sat.py
