+++ src/schedlib/policies/__init__.py
@@ -1,3 +1,2 @@
-from ..core import BasePolicy
-from .flex import FlexPolicy
 from .sat import SATPolicy
+from .satp1 import SATP1Policy
